Start all containers in parallel
