+++ codecov.yml
@@ -1,3 +1,3 @@
 ignore:
-  - "test/itest_utils.go"
+  - "test/*.go"
   - "commands/actions/prompter_stub.go"
